Add service functions
