+++ frontend/src/pages/Details.js
@@ -1,4 +1,4 @@
-import { useState, useEffect } from 'react'
+import { useState } from 'react'
 import styled from 'styled-components/macro'
 import getFromApi from '../lib/getFromApi'
 import PropTypes from 'prop-types'
